refactor string expression serde
